input_rc: allow command from multiple input_rc instances
Before this commit, if there were multiple instances of input_rc topic,
only the first one could be used.
This commit add a paramter that allows to choose to keep previous behavor
(default) or to use the other instances when no commands are recieved to
instances with best priority.

A maximum of four instances is used